MapEditorUiMgr의 enum class Modes, mode 멤버변수를  MapEditorScene으로 이동

+++ skul/skul/Ui/TextButtonUi.cpp
@@ -86,7 +86,7 @@ bool TextButtonUi::GetSelected() const
 
 void TextButtonUi::Selected()
 {
-	ChangeMode(mode);
+	ChangeMode((int)mode);
 	isSelected = true;
 	if(ShowSelected != nullptr)
 		ShowSelected();
